set spree_affiliate as real engine and added generator
