--- lib/tasks/spree_affiliate.rake
@@ -1,1 +1,0 @@
-# add custom rake tasks here
